pyproject.toml
[tool.poetry]
name = "swarm-test"
version = "0.1.0"
description = ""
authors = ["KitsonBroadhurst <kitson.broadhurst@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv= "^1.0.1"
swarm = {git = "https://github.com/openai/swarm.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
